+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.2] - 2018-02-14
+### Changed
+- Only update git databases when manifest file changes.
+- Only load manifests once per session.
+- Only update checkouts once per session.
+
+## [0.3.1] - 2018-02-14
 ### Fixed
 - Fix crash when dependencies have no manifest.